Removed biniary cache files. Reparsing log is actually faster. Temporarily made revision id ulong. Decoding commit comment characters.

+++ Insight.Shared/Model/Artifact.cs
@@ -24,7 +24,7 @@ namespace Insight.Shared.Model
 
         public string LocalPath { get; set; }
 
-        public int Revision { get; set; }
+        public ulong Revision { get; set; }
 
         public string ServerPath { get; set; }
